Extend server.cfg template
Mention some other CVARs the server owner should care about (frag limit,
allow upload, logging, hints to customize mapcycle and motd).

Also correct mp_timeleft to mp_timelimit (timeleft is in seconds and seems
to be read-only, it's rather to query the actual time life while the round
is running) and a typo.